+++ ShopMax/lib/resources/pages/shopify/product_search_page.dart
@@ -10,6 +10,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_app/app/controllers/controller.dart';
+import 'package:woosignal_shopify_api/models/response/product_search.dart';
 import '/bootstrap/helpers.dart';
 import '/resources/widgets/safearea_widget.dart';
 import '/resources/widgets/woosignal_ui.dart';